Move components to common folder
