+++ ppr-ui/src/components/common/index.ts
@@ -22,4 +22,6 @@ export { default as SharedStaffPayment } from './SharedStaffPayment.vue'
 export { default as InfoChip } from './InfoChip.vue'
 export { default as FormField } from './FormField.vue'
 export { default as WysiwygEditor } from './WysiwygEditor.vue'
+export { default as DocumentId } from './DocumentId.vue'
+export { default as Remarks } from './Remarks.vue'
 export { default as ContactInformation } from './ContactInformation.vue'
